requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
pymongo==4.8.0
openpyxl==3.1.5
reportlab==4.2.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.151.0
pandas  